src/ml_pipeline/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "This is the container for the machine learning pipeline portion of the project"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.1.0",
    "einops>=0.8.1",
    "faiss-cpu>=1.12.0",
    "google-cloud-storage>=3.4.0",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.1",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "transformers>=4.56.2",
]
